Some improvements to log & FrameBuffer + misc warning fixes

+++ src/renderer/AsyncResourceGatherer.cpp
@@ -10,6 +10,7 @@
 #include "../helpers/MiscFunctions.hpp"
 #include "../helpers/Jpeg.hpp"
 #include "../helpers/Webp.hpp"
+#include "src/helpers/Color.hpp"
 
 CAsyncResourceGatherer::CAsyncResourceGatherer() {
     if (g_pHyprlock->getScreencopy())
@@ -345,7 +346,7 @@ void CAsyncResourceGatherer::renderText(const SPreloadRequest& rq) {
     target.cairo        = CAIRO;
     target.cairosurface = CAIROSURFACE;
     target.data         = cairo_image_surface_get_data(CAIROSURFACE);
-    target.size         = {layoutWidth / PANGO_SCALE, layoutHeight / PANGO_SCALE};
+    target.size         = {layoutWidth / (double)PANGO_SCALE, layoutHeight / (double)PANGO_SCALE};
 
     std::lock_guard lg{preloadTargetsMutex};
     preloadTargets.push_back(target);
